+++ src/const.js
@@ -1,5 +1,4 @@
-// TODO: to movies.js
-export const MAX_DESCRIPTION_LENGTH = 140;
+
 
 // TODO: to comments.js
 export const EMOTIONS = [
