--- src/mock/profile-mock.js
@@ -1,7 +1,0 @@
-export const generateProfile = () => ({
-  id: 1,
-  profileTitle: 'Movie Buff',
-  avatar: 'images/bitmap@2x.png',
-});
-
-// TODO: Удалить, profileTitle вычислять по количеству просмотренных фильмов, см. ТЗ.
