--- src/utils/common/getRandomArrayElement.js
@@ -1,6 +1,0 @@
-import { getRandomInteger } from './getRandomInteger.js';
-
-export const getRandomArrayElement = (array) => array[getRandomInteger(0, array.length - 1)];
-
-// TODO: собрать функции для получения рандомных значений в random.js
-// Потом когда моки удалятся, этот модуль тоже нужно будет удалить
