--- src/utils/common/getRandomBoolean.js
@@ -1,3 +1,0 @@
-import { getRandomInteger } from './getRandomInteger.js';
-
-export const getRandomBoolean = () => Boolean(getRandomInteger(0, 1));
