--- src/utils/common/isEscEvent.js
@@ -1,3 +1,0 @@
-export const isEscEvent = (evt) => evt.key === 'Escape' || evt.key === 'Esc';
-
-// TODO: собрать подобные функции в common.js или какой-то такой общий модуль
