--- src/utils/movie/isMovieInCategory.js
@@ -1,3 +1,0 @@
-export const isMovieInCategory = (property) => Object.values(property).some(Boolean);
-
-// TODO: функция не используется
