--- src/utils/render/createElement.js
@@ -1,8 +1,0 @@
-export const createElement = (template) => {
-  const newElement = document.createElement('div');
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-// TODO: render.js
