--- src/utils/render/renderPosition.js
@@ -1,7 +1,0 @@
-export const RenderPosition = {
-  AFTERBEGIN: 'afterbegin',
-  BEFOREEND: 'beforeend',
-  AFTEREND: 'afterend',
-};
-
-// TODO: render.js
