+++ src/view/card.js
@@ -1,34 +1,29 @@
 import Abstract from './abstract.js';
-import dayjs from 'dayjs';
-import { getFormattedDuration } from '../utils/movie/getFormattedDuration.js';
-import { getFormattedDescription } from '../utils/movie/getFormattedDescription.js';
+import { getFormattedDuration, getFormattedDescription } from '../utils/movies.js';
+import { getFormattedReleaseYear } from '../utils/dates.js';
 
 const createCardTemplate = (movie) => {
   const {filmInfo, comments} = movie;
 
-  // TODO: Деструктурировать filmInfo
-  // const {title, totalRating, runtime... } = filmInfo
-  const title = filmInfo.title;
-  const rating = filmInfo.totalRating;
-  const year = dayjs(filmInfo.release.date).format('YYYY');
-  const duration = getFormattedDuration(filmInfo.runtime);
-  const genre = filmInfo.genre[0];
-  const url = filmInfo.poster;
-  const description = getFormattedDescription(filmInfo.description);
+  const {title, totalRating, runtime, genre, poster, description, release} = filmInfo;
 
-  // TODO: В каком случае может не быть поля comments?
+  const releaseYear = getFormattedReleaseYear(release.date);
+  const duration = getFormattedDuration(runtime);
+  const formattedDescription = getFormattedDescription(description);
+  const displayedGenre = genre[0];
+
   const commentsQuantity = comments ? comments.length : '0';
 
   return `<article class="film-card">
     <h3 class="film-card__title">${title}</h3>
-    <p class="film-card__rating">${rating}</p>
+    <p class="film-card__rating">${totalRating}</p>
     <p class="film-card__info">
-      <span class="film-card__year">${year}</span>
+      <span class="film-card__year">${releaseYear}</span>
       <span class="film-card__duration">${duration}</span>
-      <span class="film-card__genre">${genre}</span>
-    </p>
-    <img src=${url} alt="${title}" class="film-card__poster">
-    <p class="film-card__description">${description}</p>
+      <span class="film-card__genre">${displayedGenre}</span>
+    </p>Ааа
+    <img src=${poster} alt="${title}" class="film-card__poster">
+    <p class="film-card__description">${formattedDescription}</p>
     <a class="film-card__comments">${commentsQuantity} ${commentsQuantity === 1 ? 'comment' : 'comments'}</a>
     <div class="film-card__controls">
       <button class="film-card__controls-item film-card__controls-item--add-to-watchlist" type="button">Add to watchlist</button>
